Feat(Descriptor API): Move channel keys to the descriptor-api sub-project
