+++ channel/src/main/java/com/synopsys/integration/alert/channel/azure/boards/AzureBoardsChannel.java
@@ -44,6 +44,7 @@ import com.synopsys.integration.alert.common.event.DistributionEvent;
 import com.synopsys.integration.alert.common.event.EventManager;
 import com.synopsys.integration.alert.common.persistence.accessor.FieldUtility;
 import com.synopsys.integration.alert.common.rest.ProxyManager;
+import com.synopsys.integration.alert.descriptor.api.AzureBoardsChannelKey;
 import com.synopsys.integration.exception.IntegrationException;
 
 @Component
