+++ channel/src/main/java/com/synopsys/integration/alert/channel/azure/boards/actions/AzureBoardsConfigurationAction.java
@@ -25,8 +25,8 @@ package com.synopsys.integration.alert.channel.azure.boards.actions;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import com.synopsys.integration.alert.channel.azure.boards.AzureBoardsChannelKey;
 import com.synopsys.integration.alert.common.action.ConfigurationAction;
+import com.synopsys.integration.alert.descriptor.api.AzureBoardsChannelKey;
 
 @Component
 public class AzureBoardsConfigurationAction extends ConfigurationAction {
@@ -38,4 +38,5 @@ public class AzureBoardsConfigurationAction extends ConfigurationAction {
         addGlobalApiAction(azureBoardsGlobalApiAction);
         addDistributionTestAction(azureBoardsDistributionTestAction);
     }
+
 }
