+++ channel/src/main/java/com/synopsys/integration/alert/channel/azure/boards/oauth/storage/AzureBoardsCredentialDataStoreFactory.java
@@ -30,9 +30,9 @@ import org.springframework.stereotype.Component;
 import com.google.api.client.auth.oauth2.StoredCredential;
 import com.google.api.client.util.store.AbstractDataStoreFactory;
 import com.google.api.client.util.store.DataStore;
-import com.synopsys.integration.alert.channel.azure.boards.AzureBoardsChannelKey;
 import com.synopsys.integration.alert.common.exception.AlertRuntimeException;
 import com.synopsys.integration.alert.common.persistence.accessor.ConfigurationAccessor;
+import com.synopsys.integration.alert.descriptor.api.AzureBoardsChannelKey;
 
 @Component
 public class AzureBoardsCredentialDataStoreFactory extends AbstractDataStoreFactory {
