+++ channel/src/main/java/com/synopsys/integration/alert/channel/email/actions/EmailConfigurationAction.java
@@ -25,8 +25,8 @@ package com.synopsys.integration.alert.channel.email.actions;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import com.synopsys.integration.alert.channel.email.EmailChannelKey;
 import com.synopsys.integration.alert.common.action.ConfigurationAction;
+import com.synopsys.integration.alert.descriptor.api.EmailChannelKey;
 
 @Component
 public class EmailConfigurationAction extends ConfigurationAction {
