+++ channel/src/main/java/com/synopsys/integration/alert/channel/msteams/descriptor/MsTeamsUIConfig.java
@@ -27,10 +27,10 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import com.synopsys.integration.alert.channel.msteams.MsTeamsKey;
 import com.synopsys.integration.alert.common.descriptor.config.field.ConfigField;
 import com.synopsys.integration.alert.common.descriptor.config.field.URLInputConfigField;
 import com.synopsys.integration.alert.common.descriptor.config.ui.ChannelDistributionUIConfig;
+import com.synopsys.integration.alert.descriptor.api.MsTeamsKey;
 
 @Component
 public class MsTeamsUIConfig extends ChannelDistributionUIConfig {
