+++ channel/src/main/java/com/synopsys/integration/alert/channel/slack/descriptor/SlackDescriptor.java
@@ -25,8 +25,8 @@ package com.synopsys.integration.alert.channel.slack.descriptor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import com.synopsys.integration.alert.channel.slack.SlackChannelKey;
 import com.synopsys.integration.alert.common.descriptor.ChannelDescriptor;
+import com.synopsys.integration.alert.descriptor.api.SlackChannelKey;
 
 @Component
 public class SlackDescriptor extends ChannelDescriptor {
