+++ channel/src/main/java/com/synopsys/integration/alert/channel/slack/descriptor/SlackUIConfig.java
@@ -27,11 +27,11 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import com.synopsys.integration.alert.channel.slack.SlackChannelKey;
 import com.synopsys.integration.alert.common.descriptor.config.field.ConfigField;
 import com.synopsys.integration.alert.common.descriptor.config.field.TextInputConfigField;
 import com.synopsys.integration.alert.common.descriptor.config.field.URLInputConfigField;
 import com.synopsys.integration.alert.common.descriptor.config.ui.ChannelDistributionUIConfig;
+import com.synopsys.integration.alert.descriptor.api.SlackChannelKey;
 
 @Component
 public class SlackUIConfig extends ChannelDistributionUIConfig {
