+++ channel/src/test/java/com/synopsys/integration/alert/channel/jira/cloud/JiraCloudTestActionTest.java
@@ -31,6 +31,7 @@ import com.synopsys.integration.alert.common.channel.issuetracker.message.IssueS
 import com.synopsys.integration.alert.common.channel.issuetracker.message.IssueTrackerRequest;
 import com.synopsys.integration.alert.common.channel.issuetracker.service.TestIssueRequestCreator;
 import com.synopsys.integration.alert.common.message.model.MessageResult;
+import com.synopsys.integration.alert.descriptor.api.JiraCloudChannelKey;
 import com.synopsys.integration.jira.common.cloud.model.IssueSearchResponseModel;
 import com.synopsys.integration.jira.common.cloud.service.IssueSearchService;
 import com.synopsys.integration.jira.common.cloud.service.IssueService;
@@ -192,4 +193,5 @@ public class JiraCloudTestActionTest {
         issueResponse.getTransitions().addAll(ids);
         return issueResponse;
     }
+
 }
