+++ channel/src/test/java/com/synopsys/integration/alert/channel/jira/server/JiraServerRequestDelegatorTest.java
@@ -32,6 +32,7 @@ import com.synopsys.integration.alert.common.channel.issuetracker.message.IssueS
 import com.synopsys.integration.alert.common.channel.issuetracker.message.IssueTrackerIssueResponseModel;
 import com.synopsys.integration.alert.common.channel.issuetracker.message.IssueTrackerRequest;
 import com.synopsys.integration.alert.common.channel.issuetracker.message.IssueTrackerResponse;
+import com.synopsys.integration.alert.descriptor.api.JiraServerChannelKey;
 import com.synopsys.integration.jira.common.model.components.IdComponent;
 import com.synopsys.integration.jira.common.model.components.ProjectComponent;
 import com.synopsys.integration.jira.common.model.components.StatusDetailsComponent;
@@ -283,4 +284,5 @@ public class JiraServerRequestDelegatorTest {
         issueResponse.getTransitions().addAll(ids);
         return issueResponse;
     }
+
 }
