+++ descriptor-api/src/main/java/com/synopsys/integration/alert/descriptor/api/AzureBoardsChannelKey.java
@@ -20,16 +20,15 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.alert.channel.email;
+package com.synopsys.integration.alert.descriptor.api;
 
 import org.springframework.stereotype.Component;
 
-import com.synopsys.integration.alert.channel.email.descriptor.EmailDescriptor;
-import com.synopsys.integration.alert.descriptor.api.model.ChannelKey;
+import com.synopsys.integration.alert.descriptor.api.model.IssueTrackerChannelKey;
 
 @Component
-public final class EmailChannelKey extends ChannelKey {
-    private static final String COMPONENT_NAME = "channel_email";
+public class AzureBoardsChannelKey extends IssueTrackerChannelKey {
+    private static final String COMPONENT_NAME = "channel_azure_boards";
 
     @Override
     public String getUniversalKey() {
@@ -38,7 +37,7 @@ public final class EmailChannelKey extends ChannelKey {
 
     @Override
     public String getDisplayName() {
-        return EmailDescriptor.EMAIL_LABEL;
+        return "Azure Boards";
     }
 
 }
